+++ CHANGES.rst
@@ -5,6 +5,11 @@ The 1.7 series are stable versions, while the 1.8 are unstable.
 Changes displayed as "unreleased" in th stable series are released
 with any higher released unstable versions.
 
+1.8.4 (unreleased)
+------------------
+- launchpad #1249566: [bzr] avoid outgoing requests in offline mode
+  due to lp: locations resolution
+
 1.8.3 (unreleased)
 ------------------
 - launchpad #1281262: new ``openerp-log-level`` modifier in ``openerp_scripts``